Move some tests to match the test file they belong to
